Упорядочил импорты и убрал не нужный класс
